Change pickMove() recieve only  opponent's last move.
The goal was to simplify code in the Player classes. By sending only
one parameter, the opponent's last move, the players are spared the
task of parsing the history array. Furthermore, they do not need to
have PlayerNum as an instance variable, and thus a no-args constructor.
On the first move, the chooseCorD() method sends an empty string.

Also updated method and variable names:
- pickMove() became chooseCorD(String opponentsLastMove)
- Player interface became PDPlayer

Algorithm updated to a true round-robin where each player only plays
each other once instead of twice.