Add PlayerTester class.
PlayerTester is a utility for students to test their PDPlayer classes
before they submit them. The PlayerTester creates multiple instances
of their player class and runs a 200 iteration game of Prisoner's
Dilemma against them to see that they will not crash the PDTournament
application. Basically, it verifies that they can instantiate objects
from their class and that their chooseCorD() method will always return
a "c" or "d" when called.

Use the PlayerTester by changing line 12 to reflect the name of the
player class you are testing. It currently reads:
   PDPlayer player = new BrokenPlayer1();
Simply replace "BrokenPlayer1()" with the name of your class.

The test occurs by calling the chooseCorD method for NUMROUNDS number
of times (default is 200). Each time it sends a randomly selected "c"
or "d" as the argument for "opponentsLastMove". The default number
of instances is 10, but this could be increased to 1000 or more to
verify code whose behavior hinges upon  particular sequences in the
opponent's moves.

Also added BrokenPlayer1 class, which returns "e" instead of "c" or
"d".